Refactor countries and documents API calls
